requirements.txt
langchain==0.0.218
openai==0.27.8
chromadb==0.3.26
tiktoken==0.4.0
unstructured==0.7.10
wrapt==1.15
python-dotenv==1.0.0
discord
